requirements.txt
reflex
python-dotenv
black
isort
ruff
requests
selenium
pandas
